+++ Realm/RLMUtil.hpp
@@ -27,15 +27,13 @@
 #import <realm/util/file.hpp>
 
 namespace realm {
-    class Mixed;
+class Decimal128;
+class Mixed;
+class RealmFileException;
 }
 
 @class RLMObjectSchema;
 @class RLMProperty;
-
-namespace realm {
-    class RealmFileException;
-}
 
 __attribute__((format(NSString, 1, 2)))
 NSException *RLMException(NSString *fmt, ...);
@@ -173,6 +171,7 @@ static inline NSUInteger RLMConvertNotFound(size_t index) {
 }
 
 id RLMMixedToObjc(realm::Mixed const& value);
+realm::Decimal128 RLMObjcToDecimal128(id value);
 
 // Given a bundle identifier, return the base directory on the disk within which Realm database and support files should
 // be stored.