+++ RealmSwift/Tests/ListTests.swift
@@ -113,6 +113,8 @@ class ListTests: TestCase {
         XCTAssertFalse(obj.string.contains("foobar"))
         XCTAssertFalse(obj.data.contains(Data()))
         XCTAssertFalse(obj.date.contains(Date()))
+        XCTAssertFalse(obj.decimal.contains(Decimal128()))
+        XCTAssertFalse(obj.objectId.contains(ObjectId()))
 
         XCTAssertFalse(obj.intOpt.contains { $0 == nil })
         XCTAssertFalse(obj.int8Opt.contains { $0 == nil })
@@ -124,6 +126,8 @@ class ListTests: TestCase {
         XCTAssertFalse(obj.stringOpt.contains { $0 == nil })
         XCTAssertFalse(obj.dataOpt.contains { $0 == nil })
         XCTAssertFalse(obj.dateOpt.contains { $0 == nil })
+        XCTAssertFalse(obj.decimalOpt.contains { $0 == nil })
+        XCTAssertFalse(obj.objectIdOpt.contains { $0 == nil })
     }
 
     func testInvalidated() {
@@ -410,188 +414,101 @@ class ListTests: TestCase {
     }
 
     func testValueForKey() {
-        do {
-            let realm = try! Realm()
-            try! realm.write {
-                for value in [1, 2] {
-                    let listObject = SwiftListOfSwiftObject()
-                    let object = SwiftObject()
-                    object.intCol = value
-                    object.doubleCol = Double(value)
-                    object.stringCol = String(value)
-                    listObject.array.append(object)
-                    realm.add(listObject)
-                }
-            }
-
-            do {
-                let objects = realm.objects(SwiftListOfSwiftObject.self)
-
-                let properties: [Int] = Array(objects.flatMap { $0.array.map { $0.intCol }})
-                let listsOfObjects = objects.value(forKeyPath: "array") as! [List<SwiftObject>]
-                let kvcProperties: [Int] = Array(listsOfObjects.flatMap { $0.map { $0.intCol }})
-
-                XCTAssertEqual(properties, kvcProperties)
-            }
-            do {
-                let objects = realm.objects(SwiftListOfSwiftObject.self)
-
-                let properties = Array(objects.flatMap { $0.array.map { $0.doubleCol }})
-                let listsOfObjects = objects.value(forKeyPath: "array") as! [List<SwiftObject>]
-                let kvcProperties = Array(listsOfObjects.flatMap { $0.map { $0.doubleCol }})
-
-                XCTAssertEqual(properties, kvcProperties)
-            }
-            do {
-                let objects = realm.objects(SwiftListOfSwiftObject.self)
-
-                let properties = Array(objects.flatMap { $0.array.map { $0.stringCol }})
-                let listsOfObjects = objects.value(forKeyPath: "array") as! [List<SwiftObject>]
-                let kvcProperties = Array(listsOfObjects.flatMap { $0.map { $0.stringCol }})
-
-                XCTAssertEqual(properties, kvcProperties)
-            }
-
-            do {
-                let realm = try! Realm()
-                do {
-                    let objects = realm.objects(SwiftObject.self)
-
-                    let properties = Array(objects.compactMap { $0.intCol })
-                    let listsOfObjects = objects.value(forKeyPath: "intCol") as! [Int]
-                    let kvcProperties = Array(listsOfObjects.compactMap { $0 })
-
-                    XCTAssertEqual(properties, kvcProperties)
-                }
-                do {
-                    let objects = realm.objects(SwiftObject.self)
-
-                    let properties = Array(objects.compactMap { $0.doubleCol })
-                    let listsOfObjects = objects.value(forKeyPath: "doubleCol") as! [Double]
-                    let kvcProperties = Array(listsOfObjects.compactMap { $0 })
-
-                    XCTAssertEqual(properties, kvcProperties)
-                }
-                do {
-                    let objects = realm.objects(SwiftObject.self)
-
-                    let properties = Array(objects.compactMap { $0.stringCol })
-                    let listsOfObjects = objects.value(forKeyPath: "stringCol") as! [String]
-                    let kvcProperties = Array(listsOfObjects.compactMap { $0 })
-
-                    XCTAssertEqual(properties, kvcProperties)
-                }
+        let realm = try! Realm()
+        try! realm.write {
+            for value in [1, 2] {
+                let listObject = SwiftListOfSwiftObject()
+                let object = SwiftObject()
+                object.intCol = value
+                object.doubleCol = Double(value)
+                object.stringCol = String(value)
+                object.decimalCol = Decimal128(number: value as NSNumber)
+                object.objectIdCol = try! ObjectId(string: String(repeating: String(value), count: 24))
+                listObject.array.append(object)
+                realm.add(listObject)
             }
         }
 
-        do {
-            let realm = try! Realm()
-            try! realm.write {
-                for value in [1, 2] {
-                    let listObject = SwiftListOfSwiftOptionalObject()
-                    let object = SwiftOptionalObject()
-                    object.optIntCol.value = value
-                    object.optInt8Col.value = Int8(value)
-                    object.optDoubleCol.value = Double(value)
-                    object.optStringCol = String(value)
-                    object.optNSStringCol = NSString(format: "%d", value)
-                    listObject.array.append(object)
-                    realm.add(listObject)
-                }
-            }
+        let listObjects = realm.objects(SwiftListOfSwiftObject.self)
+        let listsOfObjects = listObjects.value(forKeyPath: "array") as! [List<SwiftObject>]
+        let objects = realm.objects(SwiftObject.self)
 
-            do {
-                let objects = realm.objects(SwiftListOfSwiftOptionalObject.self)
+        func testProperty<T: Equatable>(line: UInt = #line, fn: @escaping (SwiftObject) -> T) {
+            let properties: [T] = Array(listObjects.flatMap { $0.array.map(fn) })
+            let kvcProperties: [T] = Array(listsOfObjects.flatMap { $0.map(fn) })
+            XCTAssertEqual(properties, kvcProperties, line: line)
+        }
+        func testProperty<T: Equatable>(_ name: String, line: UInt = #line, fn: @escaping (SwiftObject) -> T) {
+            let properties = Array(objects.compactMap(fn))
+            let listsOfObjects = objects.value(forKeyPath: name) as! [T]
+            let kvcProperties = Array(listsOfObjects.compactMap { $0 })
 
-                let properties = Array(objects.flatMap { $0.array.compactMap { $0.optIntCol.value }})
-                let listsOfObjects = objects.value(forKeyPath: "array") as! [List<SwiftOptionalObject>]
-                let kvcProperties = Array(listsOfObjects.flatMap { $0.compactMap { $0.optIntCol.value }})
+            XCTAssertEqual(properties, kvcProperties, line: line)
+        }
 
-                XCTAssertEqual(properties, kvcProperties)
-            }
-            do {
-                let objects = realm.objects(SwiftListOfSwiftOptionalObject.self)
+        testProperty { $0.intCol }
+        testProperty { $0.doubleCol }
+        testProperty { $0.stringCol }
+        testProperty { $0.decimalCol }
+        testProperty { $0.objectIdCol }
 
-                let properties = Array(objects.flatMap { $0.array.compactMap { $0.optInt8Col.value }})
-                let listsOfObjects = objects.value(forKeyPath: "array") as! [List<SwiftOptionalObject>]
-                let kvcProperties = Array(listsOfObjects.flatMap { $0.compactMap { $0.optInt8Col.value }})
+        testProperty("intCol") { $0.intCol }
+        testProperty("doubleCol") { $0.doubleCol }
+        testProperty("stringCol") { $0.stringCol }
+        testProperty("decimalCol") { $0.decimalCol }
+        testProperty("objectIdCol") { $0.objectIdCol }
+    }
 
-                XCTAssertEqual(properties, kvcProperties)
-            }
-            do {
-                let objects = realm.objects(SwiftListOfSwiftOptionalObject.self)
-
-                let properties = Array(objects.flatMap { $0.array.compactMap { $0.optDoubleCol.value }})
-                let listsOfObjects = objects.value(forKeyPath: "array") as! [List<SwiftOptionalObject>]
-                let kvcProperties = Array(listsOfObjects.flatMap { $0.compactMap { $0.optDoubleCol.value }})
-
-                XCTAssertEqual(properties, kvcProperties)
-            }
-            do {
-                let objects = realm.objects(SwiftListOfSwiftOptionalObject.self)
-
-                let properties = Array(objects.flatMap { $0.array.compactMap { $0.optStringCol }})
-                let listsOfObjects = objects.value(forKeyPath: "array") as! [List<SwiftOptionalObject>]
-                let kvcProperties = Array(listsOfObjects.flatMap { $0.compactMap { $0.optStringCol }})
-
-                XCTAssertEqual(properties, kvcProperties)
-            }
-            do {
-                let objects = realm.objects(SwiftListOfSwiftOptionalObject.self)
-
-                let properties = Array(objects.flatMap { $0.array.compactMap { $0.optNSStringCol }})
-                let listsOfObjects = objects.value(forKeyPath: "array") as! [List<SwiftOptionalObject>]
-                let kvcProperties = Array(listsOfObjects.flatMap { $0.compactMap { $0.optNSStringCol }})
-
-                XCTAssertEqual(properties, kvcProperties)
-            }
-
-            do {
-                let objects = realm.objects(SwiftOptionalObject.self)
-
-                let properties = Array(objects.compactMap { $0.optIntCol.value })
-                let listsOfObjects = objects.value(forKeyPath: "optIntCol") as! [Int]
-                let kvcProperties = Array(listsOfObjects.compactMap { $0 })
-
-                XCTAssertEqual(properties, kvcProperties)
-            }
-            do {
-                let objects = realm.objects(SwiftOptionalObject.self)
-
-                let properties = Array(objects.compactMap { $0.optInt8Col.value })
-                let listsOfObjects = objects.value(forKeyPath: "optInt8Col") as! [Int8]
-                let kvcProperties = Array(listsOfObjects.compactMap { $0 })
-
-                XCTAssertEqual(properties, kvcProperties)
-            }
-            do {
-                let objects = realm.objects(SwiftOptionalObject.self)
-
-                let properties = Array(objects.compactMap { $0.optDoubleCol.value })
-                let listsOfObjects = objects.value(forKeyPath: "optDoubleCol") as! [Double]
-                let kvcProperties = Array(listsOfObjects.compactMap { $0 })
-
-                XCTAssertEqual(properties, kvcProperties)
-            }
-            do {
-                let objects = realm.objects(SwiftOptionalObject.self)
-
-                let properties = Array(objects.compactMap { $0.optStringCol })
-                let listsOfObjects = objects.value(forKeyPath: "optStringCol") as! [String]
-                let kvcProperties = Array(listsOfObjects.compactMap { $0 })
-
-                XCTAssertEqual(properties, kvcProperties)
-            }
-            do {
-                let objects = realm.objects(SwiftOptionalObject.self)
-
-                let properties = Array(objects.compactMap { $0.optNSStringCol })
-                let listsOfObjects = objects.value(forKeyPath: "optNSStringCol") as! [NSString?]
-                let kvcProperties = Array(listsOfObjects.compactMap { $0 })
-
-                XCTAssertEqual(properties, kvcProperties)
+    func testValueForKeyOptional() {
+        let realm = try! Realm()
+        try! realm.write {
+            for value in [1, 2] {
+                let listObject = SwiftListOfSwiftOptionalObject()
+                let object = SwiftOptionalObject()
+                object.optIntCol.value = value
+                object.optInt8Col.value = Int8(value)
+                object.optDoubleCol.value = Double(value)
+                object.optStringCol = String(value)
+                object.optNSStringCol = NSString(format: "%d", value)
+                object.optDecimalCol = Decimal128(number: value as NSNumber)
+                object.optObjectIdCol = try! ObjectId(string: String(repeating: String(value), count: 24))
+                listObject.array.append(object)
+                realm.add(listObject)
             }
         }
+
+        let listObjects = realm.objects(SwiftListOfSwiftOptionalObject.self)
+        let listsOfObjects = listObjects.value(forKeyPath: "array") as! [List<SwiftOptionalObject>]
+        let objects = realm.objects(SwiftOptionalObject.self)
+
+        func testProperty<T: Equatable>(line: UInt = #line, fn: @escaping (SwiftOptionalObject) -> T) {
+            let properties: [T] = Array(listObjects.flatMap { $0.array.map(fn) })
+            let kvcProperties: [T] = Array(listsOfObjects.flatMap { $0.map(fn) })
+            XCTAssertEqual(properties, kvcProperties, line: line)
+        }
+        func testProperty<T: Equatable>(_ name: String, line: UInt = #line, fn: @escaping (SwiftOptionalObject) -> T) {
+            let properties = Array(objects.compactMap(fn))
+            let listsOfObjects = objects.value(forKeyPath: name) as! [T]
+            let kvcProperties = Array(listsOfObjects.compactMap { $0 })
+
+            XCTAssertEqual(properties, kvcProperties, line: line)
+        }
+
+        testProperty { $0.optIntCol.value }
+        testProperty { $0.optInt8Col.value }
+        testProperty { $0.optDoubleCol.value }
+        testProperty { $0.optStringCol }
+        testProperty { $0.optNSStringCol }
+        testProperty { $0.optDecimalCol }
+        testProperty { $0.optObjectCol }
+
+        testProperty("optIntCol") { $0.optIntCol.value }
+        testProperty("optInt8Col") { $0.optInt8Col.value }
+        testProperty("optDoubleCol") { $0.optDoubleCol.value }
+        testProperty("optStringCol") { $0.optStringCol }
+        testProperty("optNSStringCol") { $0.optNSStringCol }
+        testProperty("optDecimalCol") { $0.optDecimalCol }
+        testProperty("optObjectCol") { $0.optObjectCol }
     }
 }