+++ RealmSwift/Tests/PrimitiveListTests.swift
@@ -35,8 +35,23 @@ struct UnmanagedObjectFactory: ObjectFactory {
 protocol ValueFactory {
     associatedtype T: RealmCollectionValue
     associatedtype W: RealmCollectionValue = T
+    associatedtype AverageType: AddableType = Double
     static func array(_ obj: SwiftListObject) -> List<T>
     static func values() -> [T]
+    static func doubleValue(_ value: AverageType) -> Double
+    static func doubleValue(t value: T) -> Double
+    static func doubleValue(w value: W) -> Double
+}
+extension ValueFactory {
+    static func doubleValue(_ value: Double) -> Double {
+        return value
+    }
+    static func doubleValue(t value: T) -> Double {
+        return (value as! NSNumber).doubleValue
+    }
+    static func doubleValue(w value: W) -> Double {
+        return (value as! NSNumber).doubleValue
+    }
 }
 
 struct IntFactory: ValueFactory {
@@ -136,6 +151,39 @@ struct DateFactory: ValueFactory {
 
     static func values() -> [Date] {
         return [Date(), Date().addingTimeInterval(10), Date().addingTimeInterval(20)]
+    }
+}
+
+struct DecimalFactory: ValueFactory {
+    typealias AverageType = Decimal128
+
+    static func array(_ obj: SwiftListObject) -> List<Decimal128> {
+        return obj.decimal
+    }
+
+    static func values() -> [Decimal128] {
+        return [Decimal128(number: 1), Decimal128(number: 2), Decimal128(number: 3)]
+    }
+
+    static func doubleValue(_ value: Decimal128) -> Double {
+        return value.doubleValue
+    }
+    static func doubleValue(t value: Decimal128) -> Double {
+        return value.doubleValue
+    }
+    static func doubleValue(w value: Decimal128) -> Double {
+        return value.doubleValue
+    }
+}
+
+struct ObjectIdFactory: ValueFactory {
+    static func array(_ obj: SwiftListObject) -> List<ObjectId> {
+        return obj.objectId
+    }
+
+    static private let _values = [ObjectId.generate(), ObjectId.generate(), ObjectId.generate()]
+    static func values() -> [ObjectId] {
+        return _values
     }
 }
 
@@ -256,6 +304,41 @@ struct OptionalDateFactory: ValueFactory {
 
     static func values() -> [Date?] {
         return [nil, Date(), Date().addingTimeInterval(20)]
+    }
+}
+
+struct OptionalDecimalFactory: ValueFactory {
+    typealias W = Decimal128
+    typealias AverageType = Decimal128
+
+    static func array(_ obj: SwiftListObject) -> List<Decimal128?> {
+        return obj.decimalOpt
+    }
+
+    static func values() -> [Decimal128?] {
+        return [nil] + DecimalFactory.values().dropLast()
+    }
+
+    static func doubleValue(_ value: Decimal128) -> Double {
+        return value.doubleValue
+    }
+    static func doubleValue(t value: Decimal128?) -> Double {
+        return value!.doubleValue
+    }
+    static func doubleValue(w value: Decimal128) -> Double {
+        return value.doubleValue
+    }
+}
+
+struct OptionalObjectIdFactory: ValueFactory {
+    typealias W = ObjectId
+
+    static func array(_ obj: SwiftListObject) -> List<ObjectId?> {
+        return obj.objectIdOpt
+    }
+
+    static func values() -> [ObjectId?] {
+        return [nil] + ObjectIdFactory.values().dropLast()
     }
 }
 
@@ -542,16 +625,15 @@ class AddablePrimitiveListTests<O: ObjectFactory, V: ValueFactory>: PrimitiveLis
         // Expressing "can be added and converted to a floating point type" as
         // a protocol requirement is awful, so sidestep it all with obj-c
         let expected = ((values.map(dynamicBridgeCast) as NSArray).value(forKeyPath: "@sum.self")! as! NSNumber).doubleValue
-        let actual: V.T = array.sum()
-        XCTAssertEqual((dynamicBridgeCast(fromSwift: actual) as! NSNumber).doubleValue, expected, accuracy: 0.01)
+        XCTAssertEqual(V.doubleValue(t: array.sum()), expected, accuracy: 0.01)
     }
 
     func testAverage() {
-        XCTAssertNil(array.average())
+        XCTAssertNil(array.average() as V.AverageType?)
         array.append(objectsIn: values)
 
         let expected = ((values.map(dynamicBridgeCast) as NSArray).value(forKeyPath: "@avg.self")! as! NSNumber).doubleValue
-        XCTAssertEqual(array.average()!, expected, accuracy: 0.01)
+        XCTAssertEqual(V.doubleValue(array.average()!), expected, accuracy: 0.01)
     }
 }
 
@@ -572,19 +654,18 @@ class OptionalAddablePrimitiveListTests<O: ObjectFactory, V: ValueFactory>: Prim
         // Expressing "can be added and converted to a floating point type" as
         // a protocol requirement is awful, so sidestep it all with obj-c
         let expected = ((nonNil.map(dynamicBridgeCast) as NSArray).value(forKeyPath: "@sum.self")! as! NSNumber).doubleValue
-        let actual: V.W = array2.sum()
-        XCTAssertEqual((dynamicBridgeCast(fromSwift: actual) as! NSNumber).doubleValue, expected, accuracy: 0.01)
+        XCTAssertEqual(V.doubleValue(w: array2.sum()), expected, accuracy: 0.01)
     }
 
     func testAverage() {
-        XCTAssertNil(array2.average())
+        XCTAssertNil(array2.average() as Double?)
         array.append(objectsIn: values)
 
         var nonNil = values!
         nonNil.remove(at: 0)
 
         let expected = ((nonNil.map(dynamicBridgeCast) as NSArray).value(forKeyPath: "@avg.self")! as! NSNumber).doubleValue
-        XCTAssertEqual(array2.average()!, expected, accuracy: 0.01)
+        XCTAssertEqual(V.doubleValue(array2.average()!), expected, accuracy: 0.01)
     }
 }
 
@@ -625,6 +706,8 @@ func addTests<OF: ObjectFactory>(_ suite: XCTestSuite, _ type: OF.Type) {
     _ = PrimitiveListTests<OF, StringFactory>._defaultTestSuite().tests.map(suite.addTest)
     _ = PrimitiveListTests<OF, DataFactory>._defaultTestSuite().tests.map(suite.addTest)
     _ = PrimitiveListTests<OF, DateFactory>._defaultTestSuite().tests.map(suite.addTest)
+    _ = PrimitiveListTests<OF, DecimalFactory>._defaultTestSuite().tests.map(suite.addTest)
+    _ = PrimitiveListTests<OF, ObjectIdFactory>._defaultTestSuite().tests.map(suite.addTest)
 
     _ = MinMaxPrimitiveListTests<OF, IntFactory>._defaultTestSuite().tests.map(suite.addTest)
     _ = MinMaxPrimitiveListTests<OF, Int8Factory>._defaultTestSuite().tests.map(suite.addTest)
@@ -634,6 +717,7 @@ func addTests<OF: ObjectFactory>(_ suite: XCTestSuite, _ type: OF.Type) {
     _ = MinMaxPrimitiveListTests<OF, FloatFactory>._defaultTestSuite().tests.map(suite.addTest)
     _ = MinMaxPrimitiveListTests<OF, DoubleFactory>._defaultTestSuite().tests.map(suite.addTest)
     _ = MinMaxPrimitiveListTests<OF, DateFactory>._defaultTestSuite().tests.map(suite.addTest)
+    _ = MinMaxPrimitiveListTests<OF, DecimalFactory>._defaultTestSuite().tests.map(suite.addTest)
 
     _ = AddablePrimitiveListTests<OF, IntFactory>._defaultTestSuite().tests.map(suite.addTest)
     _ = AddablePrimitiveListTests<OF, Int8Factory>._defaultTestSuite().tests.map(suite.addTest)
@@ -642,6 +726,7 @@ func addTests<OF: ObjectFactory>(_ suite: XCTestSuite, _ type: OF.Type) {
     _ = AddablePrimitiveListTests<OF, Int64Factory>._defaultTestSuite().tests.map(suite.addTest)
     _ = AddablePrimitiveListTests<OF, FloatFactory>._defaultTestSuite().tests.map(suite.addTest)
     _ = AddablePrimitiveListTests<OF, DoubleFactory>._defaultTestSuite().tests.map(suite.addTest)
+    _ = AddablePrimitiveListTests<OF, DecimalFactory>._defaultTestSuite().tests.map(suite.addTest)
 
     _ = PrimitiveListTests<OF, OptionalIntFactory>._defaultTestSuite().tests.map(suite.addTest)
     _ = PrimitiveListTests<OF, OptionalInt8Factory>._defaultTestSuite().tests.map(suite.addTest)
@@ -653,6 +738,8 @@ func addTests<OF: ObjectFactory>(_ suite: XCTestSuite, _ type: OF.Type) {
     _ = PrimitiveListTests<OF, OptionalStringFactory>._defaultTestSuite().tests.map(suite.addTest)
     _ = PrimitiveListTests<OF, OptionalDataFactory>._defaultTestSuite().tests.map(suite.addTest)
     _ = PrimitiveListTests<OF, OptionalDateFactory>._defaultTestSuite().tests.map(suite.addTest)
+    _ = PrimitiveListTests<OF, OptionalDecimalFactory>._defaultTestSuite().tests.map(suite.addTest)
+    _ = PrimitiveListTests<OF, OptionalObjectIdFactory>._defaultTestSuite().tests.map(suite.addTest)
 
     _ = OptionalMinMaxPrimitiveListTests<OF, OptionalIntFactory>._defaultTestSuite().tests.map(suite.addTest)
     _ = OptionalMinMaxPrimitiveListTests<OF, OptionalInt8Factory>._defaultTestSuite().tests.map(suite.addTest)
@@ -662,6 +749,7 @@ func addTests<OF: ObjectFactory>(_ suite: XCTestSuite, _ type: OF.Type) {
     _ = OptionalMinMaxPrimitiveListTests<OF, OptionalFloatFactory>._defaultTestSuite().tests.map(suite.addTest)
     _ = OptionalMinMaxPrimitiveListTests<OF, OptionalDoubleFactory>._defaultTestSuite().tests.map(suite.addTest)
     _ = OptionalMinMaxPrimitiveListTests<OF, OptionalDateFactory>._defaultTestSuite().tests.map(suite.addTest)
+    _ = OptionalMinMaxPrimitiveListTests<OF, OptionalDecimalFactory>._defaultTestSuite().tests.map(suite.addTest)
 
     _ = OptionalAddablePrimitiveListTests<OF, OptionalIntFactory>._defaultTestSuite().tests.map(suite.addTest)
     _ = OptionalAddablePrimitiveListTests<OF, OptionalInt8Factory>._defaultTestSuite().tests.map(suite.addTest)
@@ -670,6 +758,7 @@ func addTests<OF: ObjectFactory>(_ suite: XCTestSuite, _ type: OF.Type) {
     _ = OptionalAddablePrimitiveListTests<OF, OptionalInt64Factory>._defaultTestSuite().tests.map(suite.addTest)
     _ = OptionalAddablePrimitiveListTests<OF, OptionalFloatFactory>._defaultTestSuite().tests.map(suite.addTest)
     _ = OptionalAddablePrimitiveListTests<OF, OptionalDoubleFactory>._defaultTestSuite().tests.map(suite.addTest)
+    _ = OptionalAddablePrimitiveListTests<OF, OptionalDecimalFactory>._defaultTestSuite().tests.map(suite.addTest)
 }
 
 class UnmanagedPrimitiveListTests: TestCase {